+++ CHANGELOG.md
@@ -69,6 +69,8 @@
     * Change: In bootstrap, add backcompat for PHPUnit pre-v6.0
     * Fix: Update tests to handle deprecation of core's `print_inline_script()`
     * Change: Explicitly define return types for overridden methods
+    * Change: Use `assertStringNotContainsString()` instead of `assertNotContains()`
+    * Change: Use `assertStringContainsString()` instead of `assertContains()`
 * Change: Update copyright date (2025)
 * Change: Reduce number of tags defined in readme.txt